LLE Solver from vonSolms2004 complete
The method proposed by Von Solms et al. has been implemented and shown to work for the FH equation. However, after uncovering an error in mu_res from PCSAFT, it is apparent that there are other errors within the code which prevent it from finding the true LLE. Further investigation is required.